requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via httpx
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
eval-type-backport==0.2.0
    # via mistralai
google-ai-generativelanguage==0.6.9
    # via google-generativeai
google-api-core[grpc]==2.19.2
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.146.0
    # via google-generativeai
google-auth==2.34.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.8.1
    # via -r requirements.in
googleapis-common-protos==1.65.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.66.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.66.1
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.27.2
    # via mistralai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jsonpath-python==1.0.6
    # via mistralai
mistralai==1.1.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via typing-inspect
proto-plus==1.24.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.9.2
    # via
    #   google-generativeai
    #   mistralai
pydantic-core==2.23.4
    # via pydantic
pyparsing==3.1.4
    # via httplib2
python-dateutil==2.8.2
    # via mistralai
requests==2.32.3
    # via google-api-core
rsa==4.9
    # via google-auth
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
tqdm==4.66.5
    # via google-generativeai
typing-extensions==4.12.2
    # via
    #   google-generativeai
    #   pydantic
    #   pydantic-core
    #   typing-inspect
typing-inspect==0.9.0
    # via mistralai
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.3
    # via requests
